// generated/typescript/user-service/users/types.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.7.7
//   protoc               v5.29.3
// source: user-service/users/types.proto

 
import { ActionType, PermissionScope, ResourceType, TimestampFields } from "../../common/types";

export const protobufPackage = "user.v1";

export interface User {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  primaryPhoneNumber: string;
  phoneNumbers: string[];
  password: string;
  isActive: boolean;
  isEmailVerified: boolean;
  isMfaEnabled: boolean;
  roles: Role[];
  /** Use common timestamp fields */
  timestamps?: TimestampFields | undefined;
}

export interface Role {
  id: string;
  name: string;
  description: string;
  permissions: Permission[];
  timestamps?: TimestampFields | undefined;
}

export interface UserRole {
  id: string;
  userId: string;
  roleId: string;
  user?: User | undefined;
  role?: Role | undefined;
  timestamps?: TimestampFields | undefined;
}

export interface Permission {
  id: string;
  name: string;
  description: string;
  resource: ResourceType;
  action: ActionType;
  scope: PermissionScope;
  timestamps?: TimestampFields | undefined;
}

export const USER_V1_PACKAGE_NAME = "user.v1";